+++ packages/models-library/src/models_library/api_schemas_webserver/projects_metadata.py
@@ -1,8 +1,6 @@
-from typing import Literal, TypeAlias
+from typing import TypeAlias
 
-from models_library.projects_nodes_io import NodeID
-from models_library.utils.services_io import JsonSchemaDict
-from pydantic import BaseModel, Field, StrictBool, StrictFloat, StrictInt
+from pydantic import Field, StrictBool, StrictFloat, StrictInt
 
 from ..projects import ProjectID
 from ._base import InputSchema, OutputSchema
@@ -21,15 +19,3 @@ class ProjectMetadataGet(OutputSchema):
 
 class ProjectMetadataUpdate(InputSchema):
     custom: MetadataDict
-
-
-class ProjectMetadataPortGet(BaseModel):
-    key: NodeID = Field(
-        ...,
-        description="Project port's unique identifer. Same as the UUID of the associated port node",
-    )
-    kind: Literal["input", "output"]
-    content_schema: JsonSchemaDict | None = Field(
-        None,
-        description="jsonschema for the port's value. SEE https://json-schema.org/understanding-json-schema/",
-    )
